LayoutTests/imported/w3c/web-platform-tests/html/canvas/historical.any.worker-expected.txt


FAIL OffscreenCanvasRenderingContext2D.commit method is removed assert_equals: expected (undefined) undefined but got (function) function "function commit() {
    [native code]
}"
